refactor(macos): relocate journal copy + notification bridge actions to satisfy line-count ratchet
notificationJournalText and its restatement helpers move to
FloatingBarNotificationJournalCopy.swift; the three notification automation
actions move to DesktopAutomationBridge+Notifications.swift. Both oversized
hosts shrink below their origin/main baselines — no PR-body ratchet exception,
which is known to stall the beta train (#11564). boolParam widens to internal
for the extension; logic unchanged.
