+++ desktop/macos/Desktop/Sources/DesktopAutomationBridge.swift
@@ -3446,75 +3446,7 @@ final class DesktopAutomationActionRegistry {
       ]
     }
 
-    register(
-      name: "settings_notifications_snapshot",
-      summary: "Return notification settings and local permission state"
-    ) { _ in
-      async let settingsTask = APIClient.shared.getNotificationSettings()
-      let settings = try await settingsTask
-      let appState = await MainActor.run { AppState.current }
-      let hasPermission = appState?.hasNotificationPermission ?? false
-      let bannersDisabled = appState?.isNotificationBannerDisabled ?? false
-      return [
-        "schema": "enabled,frequency,frequency_label,has_permission,banners_disabled",
-        "enabled": settings.enabled ? "true" : "false",
-        "frequency": "\(settings.frequency)",
-        "frequency_label": settings.frequencyDescription,
-        "has_permission": hasPermission ? "true" : "false",
-        "banners_disabled": bannersDisabled ? "true" : "false",
-      ]
-    }
-
-    register(
-      name: "set_notification_settings",
-      summary: "Update notification settings via the real API",
-      params: ["enabled", "frequency"]
-    ) { params in
-      let enabled = params["enabled"].map { boolParam($0, default: true) }
-      let frequency = params["frequency"].flatMap { Int($0) }
-      let response = try await APIClient.shared.updateNotificationSettings(
-        enabled: enabled,
-        frequency: frequency
-      )
-      UserDefaults.standard.set(
-        response.enabled,
-        forKey: NotificationService.masterEnabledDefaultsKey)
-      UserDefaults.standard.set(
-        response.frequency,
-        forKey: NotificationService.frequencyDefaultsKey)
-      return [
-        "saved": "true",
-        "enabled": response.enabled ? "true" : "false",
-        "frequency": "\(response.frequency)",
-      ]
-    }
-
-    // The distributed com.omi.test.notification channel is machine-global: every
-    // running Omi bundle — the user's production install included — receives it,
-    // delivers it, and journals it into the real account's chat (#11943). This is
-    // the bundle-scoped alternative: it reaches only the bundle whose port and
-    // token the caller holds, through the same production sendNotification path.
-    register(
-      name: "send_test_notification",
-      summary: "Deliver a proactive test notification through the real sendNotification path (this bundle only)",
-      params: ["title", "message", "assistant_id"]
-    ) { params in
-      let title = params["title"] ?? "Insight"
-      let message = params["message"] ?? "Test notification from Omi"
-      let assistantId = params["assistant_id"] ?? "insight"
-      let posted = await MainActor.run { () -> Bool in
-        guard let ownerID = RuntimeOwnerIdentity.currentOwnerId() else { return false }
-        NotificationService.shared.sendNotification(
-          ownerID: ownerID,
-          title: title,
-          message: message,
-          assistantId: assistantId
-        )
-        return true
-      }
-      return ["posted": posted ? "true" : "false"]
-    }
-
+    registerNotificationActions()
     register(
       name: "rewind_settings_snapshot",
       summary: "Return Rewind settings retention and excluded-app counts"
@@ -3861,7 +3793,7 @@ final class DesktopAutomationActionRegistry {
 }
 
 /// Coerce a string param ("true"/"1"/"yes") into a Bool, falling back when absent.
-private func boolParam(_ raw: String?, default fallback: Bool) -> Bool {
+func boolParam(_ raw: String?, default fallback: Bool) -> Bool {
   guard let raw = raw?.trimmingCharacters(in: .whitespaces).lowercased(), !raw.isEmpty else {
     return fallback
   }
